requirements.txt
nats-py
python-dotenv